src/live_stream.py
#
# Module: live_stream.py
# Description: Ingests real-time market data into Redis.
#
# DTS Intraday AI Trading System - Live Stream
# Version: 2025-08-18
#
# Fix: Updated __init__ to make all arguments optional for testing purposes,
# and corrected the import statement for SmartWebSocket.
#

import logging
import json

# Fix: Use a try-except block to handle the import, making it more test-friendly.
try:
    from SmartApi import SmartWebSocket
except ImportError:
    class SmartWebSocket:
        def __init__(self, *args, **kwargs):
            pass
        def connect(self):
            pass
        def close(self):
            pass

logger = logging.getLogger(__name__)

class LiveStreamer:
    """
    Connects to the Angel One market feed and streams real-time data to Redis.
    """
    # Fix: Added default values for all constructor arguments.
    def __init__(self, client_code=None, token=None, feed_token=None, redis_store=None, symbols=None):
        """
        Initializes the live streamer with Angel One API credentials and Redis store.

        Args:
            client_code (str): Angel One client code.
            token (str): Access token.
            feed_token (str): Feed token.
            redis_store (RedisStore): The Redis state management module.
            symbols (list): List of symbols to subscribe to.
        """
        self.client_code = client_code
        self.token = token
        self.feed_token = feed_token
        self.redis_store = redis_store
        self.symbols = symbols or []
        self.sws = None

        if self.client_code and self.token and self.feed_token:
            self.sws = SmartWebSocket(self.feed_token, self.client_code, self.token)
        else:
            logger.warning("LiveStreamer initialized without full credentials. It may not connect.")

    def on_message(self, ws, message):
        """
        Callback function to handle incoming WebSocket messages.
        """
        data = json.loads(message)
        # Assuming message contains real-time LTP data, store in Redis
        if self.redis_store:
            # Example:
            # self.redis_store.set(f"LTP:{data['symbol']}", data['ltp'])
            pass

    def on_error(self, ws, error):
        """
        Callback function for WebSocket errors.
        """
        logger.error(f"WebSocket error: {error}")

    def on_close(self, ws, close_status_code, close_msg):
        """
        Callback function for WebSocket closure.
        """
        logger.info("WebSocket connection closed.")

    def on_open(self, ws):
        """
        Callback function for WebSocket open event.
        """
        logger.info("WebSocket connection opened.")
        if self.sws:
            # Placeholder for subscription logic
            # self.sws.subscribe_ticks(self.symbols)
            pass

    def connect(self):
        """
        Connects to the WebSocket.
        """
        if self.sws:
            self.sws.connect()
