requirements.txt
certifi==2025.7.14charset-normalizer==3.4.2idna==3.10python-dateutil==2.9.0.post0python-dotenv==1.1.1requests==2.32.4six==1.17.0smartapi-python==1.5.5urllib3==2.5.0pandaspyotpredislogzerowebsocket-clientflaskstreamlitplotlywerkzeug